m1/final_project/src/requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
fastapi==0.73.0
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
multidict==6.0.2
pycodestyle==2.8.0
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing-extensions==4.1.0
uvicorn[standard]==0.17.4
yarl==1.7.2
